minimal_adl_ethene_butadiene_KREG/requirements.txt
# Base dependencies
numpy>=1.24
PyYAML>=6.0
mlatom>=3.18
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3
scikit-learn>=1.3

# Model backend notes
# KREG backend is bundled inside MLatom and does not require torch.
# If you switch this project back to ANI/MACE/NequIP in the future,
# then you will need to install the corresponding backend packages separately.
